Make ScrollbarControlSize an enum class https://bugs.webkit.org/show_bug.cgi?id=199325
Reviewed by Sam Weinig.

Make ScrollbarControlSize an enum class, and make some RenderScrollbar methods const.

Source/WebCore:

* platform/ScrollTypes.h:
* platform/ScrollView.cpp:
(WebCore::ScrollView::createScrollbar):
* platform/ScrollbarTheme.h:
(WebCore::ScrollbarTheme::scrollbarThickness):
* platform/ios/ScrollbarThemeIOS.h:
* platform/mac/ScrollbarThemeMac.h:
* platform/mac/ScrollbarThemeMac.mm:
(WebCore::scrollbarControlSizeToNSControlSize):
(WebCore::scrollbarSizeToIndex):
(WebCore::ScrollbarThemeMac::hasButtons):
(WebCore::buttonRepaintRect):
(WebCore::ScrollbarThemeMac::backButtonRect):
(WebCore::ScrollbarThemeMac::forwardButtonRect):
(WebCore::ScrollbarThemeMac::trackRect):
* platform/mock/ScrollbarThemeMock.cpp:
(WebCore::ScrollbarThemeMock::scrollbarThickness):
* platform/mock/ScrollbarThemeMock.h:
* platform/playstation/ScrollbarThemePlayStation.h:
* platform/win/PopupMenuWin.cpp:
(WebCore::PopupMenuWin::show):
(WebCore::PopupMenuWin::calculatePositionAndSize):
* platform/win/ScrollbarThemeWin.h:
* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::createScrollbar):
* rendering/RenderScrollbar.cpp:
(WebCore::RenderScrollbar::RenderScrollbar):
(WebCore::RenderScrollbar::buttonRect const):
(WebCore::RenderScrollbar::trackRect const):
(WebCore::RenderScrollbar::trackPieceRectWithMargins const):
(WebCore::RenderScrollbar::minimumThumbLength const):
(WebCore::RenderScrollbar::opacity const):
(WebCore::RenderScrollbar::buttonRect): Deleted.
(WebCore::RenderScrollbar::trackRect): Deleted.
(WebCore::RenderScrollbar::trackPieceRectWithMargins): Deleted.
(WebCore::RenderScrollbar::minimumThumbLength): Deleted.
(WebCore::RenderScrollbar::opacity): Deleted.
* rendering/RenderScrollbar.h:
* rendering/RenderTheme.h:
(WebCore::RenderTheme::scrollbarControlSizeForPart):
* rendering/RenderThemeMac.h:

Source/WebKit:

* UIProcess/win/WebPopupMenuProxyWin.cpp:
(WebKit::WebPopupMenuProxyWin::showPopupMenu):
(WebKit::WebPopupMenuProxyWin::calculatePositionAndSize):
* WebProcess/Plugins/PDF/PDFPlugin.mm:
(WebKit::PDFPlugin::createScrollbar):


Canonical link: https://commits.webkit.org/227273@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@264525 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/platform/mock/ScrollbarThemeMock.cpp
@@ -41,7 +41,7 @@ IntRect ScrollbarThemeMock::trackRect(Scrollbar& scrollbar, bool)
 
 int ScrollbarThemeMock::scrollbarThickness(ScrollbarControlSize controlSize, ScrollbarExpansionState)
 {
-    return cScrollbarThickness[controlSize];
+    return cScrollbarThickness[static_cast<uint8_t>(controlSize)];
 }
 
 void ScrollbarThemeMock::paintTrackBackground(GraphicsContext& context, Scrollbar& scrollbar, const IntRect& trackRect)
